[MSG2-48][FIX] Backspace in recipient list was always deleting user even if input wasn't empty

+++ conversation/src/main/resources/public/ts/directives/recipientList.ts
@@ -73,7 +73,7 @@ export const recipientList = ng.directive('recipientList', () => {
             });
 
             element.find('input').on('keydown', function (e) {
-                if (e.keyCode === 8) { // BackSpace
+                if (e.keyCode === 8 && scope.searchText.length === 0) { // BackSpace
                     var nb = scope.ngModel.length;
                     if (nb > 0)
                         scope.deleteItem(scope.ngModel[nb - 1]);
